requirements.txt
Flask_APScheduler==1.12.4
Flask_Bcrypt==1.0.1
Flask_Login==0.5.0
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
pandas
pytz==2022.1
SQLAlchemy==1.4.31
WTForms
Werkzeug~=2.0.0
filelock
openpyxl
xlrd